#22 Reworked security so you can set a screen to need a pin
A password is only needed once per session, but a pin is required every
time you open the screen. If you haven't set a pin you will be prompted
for a password.
You can set a screen to be PasswordAndPin, which will use a password if
it hasn't been entered this session, after that first time it will ask
for the pin.